Fix the comments
